The seating plan can now add seats to the cart
The functionality to add seats from the seating was added. This thanks to the add cart button added, besides the go back button which allows the user to go back to the home page. The fragment also calculates the total based on how many seats the user adds, there is a max of 10 seats each time