Issue #6: Seating plan has been completed
Besides the seating plan, there were some minor modifications to the code for readability, and the creation of an apiURL variable in the api class, so if the website changes domain, it is easy to update. Also, a network security config xml file was added. This is the final commit for the issue 6 branch.